requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
websockets==12.0
python-dotenv==1.0.0
google-generativeai==0.3.2
faster-whisper==0.10.0
requests==2.31.0